requirements.txt
PyQt6>=6.5.0
groq>=0.4.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
trimesh>=4.0.0
rembg>=2.0.50
onnxruntime>=1.16.0
huggingface_hub>=0.20.0
einops>=0.7.0
omegaconf>=2.3.0
numpy>=1.24.0
requests>=2.31.0
scipy>=1.11.0
networkx>=3.1
transformers>=4.35.0
xatlas>=0.0.9
moderngl>=5.10.0
